requirements.txt
Flask>=3.0.0
numpy>=1.26.0
opencv-python>=4.10.0
python-dotenv>=1.0.0
groq>=0.11.0
supabase>=2.10.0
flask-bcrypt>=1.0.1
email-validator>=2.1.0
Pillow>=10.0.0
scikit-learn>=1.3.0
